refactor(agentic-ai): extract DocumentXmlTag record from handler
Move XML tag building, attribute escaping, and document short ID
extraction into a dedicated DocumentXmlTag record with factory methods
and toXml() serialization. Tests moved to DocumentXmlTagTest.
