+++ connectors-e2e-test/connectors-e2e-test-agentic-ai/src/test/java/io/camunda/connector/e2e/agenticai/aiagent/langchain4j/jobworker/L4JAiAgentJobWorkerMcpIntegrationTests.java
@@ -434,7 +434,7 @@ public class L4JAiAgentJobWorkerMcpIntegrationTests extends BaseL4JAiAgentJobWor
                       tc ->
                           assertThat(tc.text())
                               .isEqualTo(
-                                  "<document tool=\"MCP_A_MCP_Client___toolA\" call-id=\"img111\" document-short-id=\"%s\" />"
+                                  "<document tool-name=\"MCP_A_MCP_Client___toolA\" tool-call-id=\"img111\" document-short-id=\"%s\" />"
                                       .formatted(documentShortId)));
               assertThat(contents.get(2))
                   .isInstanceOfSatisfying(
