+++ connectors-e2e-test/connectors-e2e-test-agentic-ai/src/test/java/io/camunda/connector/e2e/agenticai/aiagent/langchain4j/outboundconnector/L4JAiAgentConnectorToolCallingTests.java
@@ -178,7 +178,7 @@ public class L4JAiAgentConnectorToolCallingTests extends BaseL4JAiAgentConnector
                       tc ->
                           assertThat(tc.text())
                               .isEqualTo(
-                                  "<document tool=\"Download_A_File\" call-id=\"aaa111\" document-short-id=\"%s\" />"
+                                  "<document tool-name=\"Download_A_File\" tool-call-id=\"aaa111\" document-short-id=\"%s\" />"
                                       .formatted(documentShortId)));
               assertDocumentContentBlock(contents.get(2), type, mimeType);
             });
